+++ src/DCSFP.Tests/DCSFP.Tests.csproj
@@ -40,139 +40,4 @@
 		<ProjectReference Include="..\NonVisuals\NonVisuals.csproj" />
 	</ItemGroup>
 
-	<ItemGroup>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\BackGroundImage_CduU_150x150.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\BackGroundImage_CduU_30x30.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\BackGroundImage_CduU_72x72.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\BitmapImageFormat_Original_150x150.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundColor_72x72_1.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundColor_72x72_2.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundColor_72x72_3.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundColor_72x72_4.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundColor_72x72_5.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundColor_72x72_6.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundColor_72x72_7.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage150_72x72_1.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage150_72x72_2.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage150_72x72_3.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage150_72x72_4.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage150_72x72_5.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage150_72x72_6.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage150_72x72_7.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage30_72x72_1.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage30_72x72_2.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage30_72x72_3.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage30_72x72_4.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage30_72x72_5.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage30_72x72_6.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage30_72x72_7.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage72_72x72_1.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage72_72x72_2.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage72_72x72_3.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage72_72x72_4.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage72_72x72_5.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage72_72x72_6.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Created_With_BackgroundImage72_72x72_7.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\EmptyRed_StreamdeckBitmap.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\EmptyRed_StreamdeckBitmap_72x72.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\EnlargedBitmapCanvas_FromGreaterThan_72x72.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\EnlargedBitmapCanvas_FromSmallerThan_72x72.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\IncompatibleForKeyBitmap.png">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\OriginalAdjusted_0.7_0.8_.0.9.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\OriginalAdjusted_0.7_0.8_0.9.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\OriginalAdjusted_1.0_4.0_1.0.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\OriginalAdjusted_1.2_2.0_.0.8.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\OriginalAdjusted_1.2_2.0_0.8.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Original_150x150.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="NonVisuals\BitmapCreatorTestsResources\Original_50x50.bmp">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-	</ItemGroup>
-
 </Project>
