+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/BackGroundImage_CduU_150x150.bmp
