+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/BackGroundImage_CduU_30x30.bmp
