+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/BackGroundImage_CduU_72x72.bmp
