+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/Created_With_BackgroundColor_72x72_5.bmp
