+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/Created_With_BackgroundImage150_72x72_2.bmp
