+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/Created_With_BackgroundImage30_72x72_6.bmp
