+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/Created_With_BackgroundImage72_72x72_2.bmp
