+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/Created_With_BackgroundImage72_72x72_3.bmp
