+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/Created_With_BackgroundImage72_72x72_7.bmp
