+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/EmptyRed_StreamdeckBitmap_72x72.bmp
