+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/EnlargedBitmapCanvas_FromGreaterThan_72x72.bmp
