+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/EnlargedBitmapCanvas_FromSmallerThan_72x72.bmp
