+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/IncompatibleForKeyBitmap.png
