+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/OriginalAdjusted_0.7_0.8_0.9.bmp
