+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/OriginalAdjusted_1.0_4.0_1.0.bmp
