+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/OriginalAdjusted_1.2_2.0_0.8.bmp
