+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/Original_150x150.bmp
