+++ src/NonVisuals.Tests/BitmapCreatorTestsResources/Original_50x50.bmp
