+++ src/NonVisuals.Tests/FlightRadio/FlightRadioSettings/Specific/FlightRadioSpecificSettingsTests.cs
@@ -1,4 +1,5 @@
 ﻿using NonVisuals.Radios.RadioSettings;
+using System;
 using Xunit;
 
 namespace NonVisuals.Tests.FlightRadio.FlightRadioSettings.Specific
