+++ src/NonVisuals.Tests/NonVisuals.Tests.csproj
@@ -1,9 +1,9 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net6.0-windows</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
+    <ImplicitUsings>disable</ImplicitUsings>
+    <Nullable>disable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
@@ -21,6 +21,129 @@
 
   <ItemGroup>
     <ProjectReference Include="..\NonVisuals\NonVisuals.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="BitmapCreatorTestsResources\BackGroundImage_CduU_150x150.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\BackGroundImage_CduU_30x30.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\BackGroundImage_CduU_72x72.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundColor_72x72_1.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundColor_72x72_2.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundColor_72x72_3.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundColor_72x72_4.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundColor_72x72_5.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundColor_72x72_6.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundColor_72x72_7.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage150_72x72_1.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage150_72x72_2.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage150_72x72_3.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage150_72x72_4.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage150_72x72_5.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage150_72x72_6.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage150_72x72_7.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage30_72x72_1.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage30_72x72_2.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage30_72x72_3.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage30_72x72_4.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage30_72x72_5.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage30_72x72_6.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage30_72x72_7.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage72_72x72_1.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage72_72x72_2.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage72_72x72_3.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage72_72x72_4.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage72_72x72_5.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage72_72x72_6.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Created_With_BackgroundImage72_72x72_7.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\EmptyRed_StreamdeckBitmap_72x72.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\EnlargedBitmapCanvas_FromGreaterThan_72x72.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\EnlargedBitmapCanvas_FromSmallerThan_72x72.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\IncompatibleForKeyBitmap.png">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\OriginalAdjusted_0.7_0.8_0.9.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\OriginalAdjusted_1.0_4.0_1.0.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\OriginalAdjusted_1.2_2.0_0.8.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Original_150x150.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="BitmapCreatorTestsResources\Original_50x50.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 
 </Project>
