+++ src/NonVisuals.Tests/Radios/RadioPanelPZ69A10CTests.cs
@@ -1,8 +1,8 @@
-﻿using System;
-using NonVisuals.Radios;
+﻿using NonVisuals.Radios;
+using System;
 using Xunit;
 
-namespace DCSFP.Tests.NonVisuals
+namespace NonVisuals.Tests.Radios
 {
     public class RadioPanelPZ69A10CTests
     {
