+++ src/NonVisuals.Tests/SwitchPanelPZ55Tests.cs
@@ -3,7 +3,7 @@ using NonVisuals.Panels.Saitek.Panels;
 using NonVisuals.Panels.Saitek.Switches;
 using Xunit;
 
-namespace DCSFP.Tests.NonVisuals
+namespace NonVisuals.Tests 
 {
     public class SwitchPanelPZ55Tests
     {
